backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
django-filter==23.5
python-decouple==3.8
Pillow
celery==5.3.4
redis==5.0.1
openpyxl>=3.1.0
pandas>=2.0.0

